Update the pseudo messages to reflect changes to fedocal
The data model of fedocal has changed quite a lot in 0.4.0, the fields
calendar_regional_meetings and calendar_multiple_meetings have been removed from
the database and will thus no longer be part of the fedmsg messages.

On the contrary the field meeting_timezone and calendar_status have been added and set to be published in the fedmsg messages.

The field calendar_manager_group was renamed to calendar_editor_group in order to
reflect more the actual function of this field.