Fix doc_utilities.
This broke when I added the new conglomerator tests.  The logic in
load_classes that returned the zero-eth element no longer worked (it
would return the 2 conglomerator tests while the hundreds of basic tests
didn't got discarded).

This new approach more precisely loads justs the tests we want (and
excludes the conglomerator tests -- we don't know what to do with them
anyways.)