Fix string comparison in the fedimg proc.
This was a really hard bug to track down.  tl;dr, "is" returns True if
you are comparing "foo" with "foo".  However, "is" returns False if you
are comparing "foo" with u"foo".

So, the test cases were passing.  Locally constructed tests by hand
would pass.  But the IRC bot would fail since it was handing unicode
strings into the ``subtitle(...)`` call.