No longer stuff package owners in msg2usernames for pkgdb.
It used to be that, for pkgdb messages, msg2usernames included many
usernames:  the actor performing the event, the subject on whom the
action was performed, and any package owners of the package.

That third bit was added as a hack to get message routing in FMN working
easily when it first launched.  I wanted to route pkgdb messages to
package owners when requests were made on their packages.

This is no longer necessary since FMN can now query pkgdb for who owns
what packages.. but furthermore it actually breaks other more
complicated filters that we're adding to the default set.

For instance, there is a rule that says "if someone does something to
one of my packages, tell me unless I did it".  The "unless I did it"
part is artificially evaluating to true, since everyone's usernames get
returned by msg2usernames.  This commit should fix that.