Add the explanatory message in the human readable format
When a request (new branch or new package) is blocked or denied, the
person doing the action must provide a message explaining why.
It is nicer to include this explanation in the human readable representation
of the femsg message as it is what the end-user of FMN, IRC will see.