Pad against a race condition with koji.
We *often* hit a race condition with koji here.

It publishes the fedmsg message before it has committed the details to its
database.  When we receive the message, we here try to query koji for more
details but... the fedmsg got to us before the db commit finished and koji says
"I know nothing."