Query only for active FAS users.
This is a follow-on to #383 that takes advantage of fedora-infra/fas#162.

It seems that the code from #383 fails now given that spammers have bloated our
FAS db.  Searching just for `'a*'` times out/fails.

By introducing ``'status': 'active',`` here, it gets things working again.
Querying for just `'a*'` took about 3.5 minutes on a local test at the time of
writing.