Substantial rewrite + extension of compose.* tests
The tests for the 'compose' message topic were missing rather
a lot of coverage. This commit substantially revises and
extends the tests such that they should hopefully cover just
about all message forms that have existed. This commit does not
change the behaviour of the processor at all, and so some of the
expected values are sort of 'wrong' - they represent what the
processor currently does, even though the newly-added test makes
it clear the processor should probably do something else.
Subsequent commits will improve the processor behaviour, and of
course update the tests where appropriate.

Signed-off-by: Adam Williamson <awilliam@redhat.com>