Merge the Allpackages processor into the Atomic CI processor
By itself this is not ideal, we are using the same processor for messages coming
from two different CI pipelines running in the ci.centos.org infrastructure.
However, we do not have really the choice, the first pipeline deployed there
is sending org.centos.prod/stage.ci.pipeline.<action> while the allpackages
pipeline is now sending messages of the type:
org.centos.prod/stage.ci.pipeline.allpackages and this is caught by the
Atomic CI processor.

So we had two ways to fix this:
- adjust the topic sent by the Atomic CI pipeline to include something like
  atomic_ci where the other pipeline is using 'allpackages'
- hack the processor to support both the atomic_ci and the allpackages format

Since we aim at being backward compatible, the second step was needed anyway
to support old and potentially new format of the messages sent by the atomic_ci
pipeline.
So since we do the second fix anyway, it is just as easy to not adjust the topic
sent by the Atomic CI pipeline.

Signed-off-by: Pierre-Yves Chibon <pingou@pingoured.fr>