Fix CI pipeline processor for current messages (#495) (#484)
The CI pipeline processor has an outright bug causing packagers
to be sent lots of error messages (#495), and aside from that
could not process most current messages, as it was not updated
for the introduction of separate 'allpackages-build' and
'allpackages-pr' messages (#484).

This fixes all cases where something we look at might be set to
None, and revises the tests heavily to test all current message
topics that I could find, and mark all old ones as legacy. The
tests use real messages with a couple of lengthy values taken
out and all the usernames changed to 'jankratochvil' as it makes
things easier.

Signed-off-by: Adam Williamson <awilliam@redhat.com>