bodhi: handle updates not always having a 'submitter'
We expect the update dict to have a 'submitter' key in several
places, which is the username of the submitter. However, in a
lot of messages the update dict does not have this 'submitter'
key. I'm not 100% sure of the details, but 'update.comment'
messages at least since 2019-03-06 don't have it, and
'update.edit' messages from before Bodhi 4.x *do* have it but
those since Bodhi 4.0 was deployed do *not* have it.

All messages that don't have it seem to have a 'user' in the
'update' dict instead, and this also represents the submitter
(but it's a dict with various values in it). Everywhere we were
only looking for the 'submitter' key, let's also look for the
'user' (and get out the 'name' value from the dict).

This fixes the titles for 'update.complete' messages, and
probably also for 'update.eject' messages (though none of those
have been emitted since the Bodhi 4.x deployment so I can't
check that for sure). It also changes the behaviour of a couple
of pre-existing tests, but I think the change is correct.

Signed-off-by: Adam Williamson <awilliam@redhat.com>