Bodhi: refactor objects, packages and link; don't assume title
The `title` key in the update dict is not part of the new Bodhi
message schema, so this should not assume it exists. The builds
dict *is* a part of the schema, so if we don't find the title,
we can create one based on the builds dict.

This also refactors the objects, packages and link methods to
reduce the use of huge forests of `elif`s.

Signed-off-by: Adam Williamson <awilliam@redhat.com>