Bodhi: handle modular updates.fedora.sync messages, add tests
These messages are slightly odd - despite being under the Bodhi
topic they're not published by Bodhi itself but by a script
that the Fedora infra ansible plays deploy to the Bodhi backend
server. Explain that a bit better, and improve the handling of
these messages to deal with the messages for modular updates
repo composes - we need to spot these by the 'm' suffix to the
release, and tweak the subtitle and link when we are dealing
with a modular compose. Also add tests for these messages.

Signed-off-by: Adam Williamson <awilliam@redhat.com>