Handle newer bodhi.masher.start and bodhi.composer.start
In Bodhi 3.2.0, the format of the bodhi.masher.start message
was changed significantly; the meta processor assumes these
messages will have a `updates` list, but they no longer do. So
processing of them is broken. Also in Bodhi 4, the message is
renamed to bodhi.composer.start. This handles both the newer
masher.start messages and the composer.start messages, and adds
tests. Sadly we can't actually get much useful information out
of the new form of the message: I have filed
https://github.com/fedora-infra/bodhi/issues/3360 on this.

Signed-off-by: Adam Williamson <awilliam@redhat.com>